+++ clients/javascript/lib/index.ts
@@ -116,5 +116,5 @@ export const NitteiClient = async (
 // Enums
 export { RRuleFrequency } from './gen_types/RRuleFrequency'
 export { ScheduleRuleVariant } from './types/ScheduleRuleVariant'
-export { Weekday } from './types/Weekdday'
+export { Weekday } from './types/Weekday'
 export { Frequency } from './types/Frequency'
