+++ clients/javascript/lib/types/Weekday.ts
